third_party/WebKit/LayoutTests/media/video-play-require-user-gesture-expected.txt

Test that video play() does not work unless a user clicked on the play button.

EVENT(canplaythrough)

* No user gesture initiated
RUN(video.play())
EXPECTED (video.paused == 'true') OK

* User gesture initiated
EVENT(playing)
RUN(video.pause())
EVENT(pause)
EXPECTED (video.paused == 'true') OK
END OF TEST
